refactor(organization): Change package to `com.github.displug.displug`
Modified:
- build.gradle
- Category
- Command
- CommandContext
- CommandExecuted
- CommandManager
- CommandRegister
- CommandResult
- Configuration
- Displug
- DisplugImpl
- Displugin
- DManager
- EventListener
- EventManager
- ExitCode
- Indentable
- Main
- Permissible
- PluginException
- PluginLoaded
- PluginManager
- PluginStarted
- PluginStopped
- SManager

BREAKING CHANGE: All previous package location are changed and cannot be used, please update your import ine your project
